+++ tools/oicana_cli/src/compile/export.rs
@@ -1,11 +1,11 @@
 use anyhow::{bail, Context};
 use clap::ValueEnum;
+use oicana::export::pdf::export_merged_pdf;
+use oicana::export::png::export_merged_png;
+use oicana::export::svg::export_merged_svg;
+use oicana::files::native::NativeTemplate;
+use oicana::template::PdfStandard;
 use oicana::Template;
-use oicana_export::pdf::export_merged_pdf;
-use oicana_export::png::export_merged_png;
-use oicana_export::svg::export_merged_svg;
-use oicana_files::native::NativeTemplate;
-use oicana_template::PdfStandard;
 use std::fs;
 use std::path::Path;
 use typst::layout::PagedDocument;
@@ -19,7 +19,7 @@ pub fn export_pdf(
     let standards = if let Some(cli_stds) = cli_standards {
         parse_pdf_standards(&cli_stds)?
     } else {
-        world.manifest().tool.oicana.export.pdf.standards.clone()
+        world.manifest().pdf_standards().to_vec()
     };
 
     let pdf_buffer = match export_merged_pdf(document, world, &standards) {